% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanCSVFile.R
\name{CleanCSVFile}
\alias{CleanCSVFile}
\title{Separating CSV Dataset Into four columns}
\usage{
CleanCSVFile(filepath, fluidLevels, cartThick, sevLevels, age)
}
\arguments{
\item{filepath}{A character string of the filepath to find your dataset.}

\item{fluidLevels}{A character string of the column representing in the
fluid level in data set provided.}

\item{cartThick}{A character string of the column representing in the
cartilage thickness in data set provided.}

\item{sevLevels}{A character string of the column representing in the
severity levels in data set provided.}

\item{age}{A character string of the column representing in the age of the
individuals in data set provided.}
}
\value{
Returns a data frame with four columns:
   (1) fluid levels
   (2) cartilage thickness
   (3) severity levels
   (4) patient age
}
\description{
The cleaned dataset will formated to contain 4 columns:  fluid levels,
cartilage thickness, severity levels, and patient age. In addition, each
entry of the data set will contain numeric information(i.e no NA).The return
value of this function should be a data frame with rows representing the
individuals and columns representing the four variables mentioned above.
}
\examples{
\dontrun{

exampleDataPath <- system.file("extdata",
                               "example_data.csv",
                               package = "OsteoAnalizer")

dataList <- OsteoAnalizer::CleanCSVFile(filepath = exampleDataPath,
                               fluidLevels = "Fluid Levels",
                               cartThick = "Cartilage Thickness",
                               sevLevels = "Severity Levels",
                               age = "Age")

}

}
\references{
R Core Team (2022). R: A language and environment for statistical
  computing. R Foundation for Statistical Computing, Vienna, Austria.
  URL https://www.R-project.org/.
}
\author{
{Krutika Joshi, \email{krutikadilip.joshi@mail.utoronto.ca}}
}
