% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeBox.R
\name{makeBox}
\alias{makeBox}
\title{Visualizing severity level against the fluid levels, cartilage thickness
or patient age.}
\usage{
makeBox(dataMade)
}
\arguments{
\item{dataMade}{A list created by separateCSV.}
}
\value{
Returns a 3 box plots showcasing the distribution of fluid levels,
cartilage thickness and patient age
}
\description{
The function will take in a data set and a continuous variable
(i.e fluid levels, cartilage thickness or patient age). Then the data set
will be grouped based on the severity levels and 5 box plots will be created
to examine the distribution of the parameter variable across the severity
levels. The purpose of this function is to examine the distribution of the
variable across the severity levels.
}
\examples{
\dontrun{

exampleDataPath <- system.file("extdata",
                               "example_data.csv",
                               package = "OsteoAnalizer")

dataMade <- OsteoAnalizer::CleanCSVFile(filepath = exampleDataPath,
                               fluidLevels = "fluid_levels",
                               cartThick = "cartilage_thickness",
                               sevLevels = "severity_levels",
                               age = "age")
OsteoAnalizer::makeBox(dataMade)
}

}
\references{
R Core Team (2022). R: A language and environment for statistical
  computing. R Foundation for Statistical Computing, Vienna, Austria.
  URL https://www.R-project.org/.
}
\author{
{Krutika Joshi, \email{krutikadilip.joshi@mail.utoronto.ca}}
}
