% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeScatter.R
\name{makeScatter}
\alias{makeScatter}
\title{Visualizing fluid levels, cartilage thickness or patient age by using a
scatter plot}
\usage{
makeScatter(dataMade, varOne, varTwo, reg = TRUE)
}
\arguments{
\item{dataMade}{A character string of the filepath to find your dataset.}

\item{varOne}{A character string of the first variable that needs
to be examined.}

\item{varTwo}{A character string of the second variable that needs
to be examined.}

\item{reg}{A boolean that determines if the scatter plot will contain a line
of regression.}
}
\value{
Returns a scatter plot with the two variables inputted by the user
}
\description{
This function will take in a data set, two continuous variables and a
boolean. And the output will be a scatter plot. The x-axis of the scatter
plot will contain one of the parameter variables and the y-axis will
contain the latter parameter variable. If the boolean parameter is set to
true, the scatterplot will further contain colors based on the severity
levels. This function will help understand the relation between the two
chosen variables.
}
\examples{
\dontrun{

exampleDataPath <- system.file("extdata",
                               "example_data.csv",
                               package = "OsteoAnalizer")

dataMade <- OsteoAnalizer::CleanCSVFile(filepath = exampleDataPath,
                               fluidLevels = "fluid_levels",
                               cartThick = "cartilage_thickness",
                               sevLevels = "severity_levels",
                               age = "age")

OsteoAnalizer::makeScatter(dataMade)

}

}
\references{
R Core Team (2022). R: A language and environment for statistical
  computing. R Foundation for Statistical Computing, Vienna, Austria.
  URL https://www.R-project.org/.
}
\author{
{Krutika Joshi, \email{krutikadilip.joshi@mail.utoronto.ca}}
}
