% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{my_data}
\alias{my_data}
\title{Example dataset: Mock data set}
\format{
A data frame with 36 rows and 4 columns.
}
\source{
Provide information about the source of the data, if applicable.
}
\usage{
data(my_data)
}
\description{
This dataset, `my_data`, contains information about fluid levels, cartilage
thickness, severity levels, and patient age.
}
\examples{
data(my_data)
}
\keyword{datasets}
