% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanCSVFile.R
\name{CleanCSVFile}
\alias{CleanCSVFile}
\title{Separating CSV Dataset Into four columns}
\usage{
CleanCSVFile(filepath, fluidLevels, cartThick, sevLevels, age)
}
\arguments{
\item{filepath}{A character string of the filepath to find your dataset.}

\item{fluidLevels}{A character string of the column representing in the
fluid level in data set provided.}

\item{cartThick}{A character string of the column representing in the
cartilage thickness in data set provided.}

\item{sevLevels}{A character string of the column representing in the
severity levels in data set provided.}

\item{age}{A character string of the column representing in the age of the
individuals in data set provided.}
}
\value{
Returns a list with four columns:
   (1) fluid levels
   (2) cartilage thickness
   (3) severity levels
   (4) patient age
}
\description{
The cleaned dataset will formated to contain 4 columns:  fluid levels,
cartilage thickness, severity levels, and patient age. In addition, each
entry of the data set will contain numeric information(i.e no NA).The return
value of this function should be a data frame with rows representing the
individuals and columns representing the four variables mentioned above.
}
\examples{
\dontrun{

exampleDataPath <- system.file("extdata",
                               "example_data.csv",
                               package = "OsteoAnalizer")

dataList <- OsteoAnalizer::CleanCSVFile(filepath = exampleDataPath,
                               fluidLevels = "fluid_levels",
                               cartThick = "cartilage_thickness",
                               sevLevels = "severity_levels",
                               age = "patient_age")

}

}
\references{
Chatgpt. ChatGPT. (n.d.). https://openai.com/chatgpt

Bates, C. (2023, April 9). Tutorial: Loading and cleaning data with R and
the Tidyverse. Dataquest.
https://www.dataquest.io/blog/load-clean-data-r-tidyverse/

Li, F. (2022) sexDisaggregate:To Obtain and Analyze Sex-disaggregated Datasets,
Unpublished. URL https://github.com/lifangy6/sexDisaggregate"
}
\author{
{Krutika Joshi, \email{krutikadilip.joshi@mail.utoronto.ca}}
}
