% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeBox.R
\name{makeBox}
\alias{makeBox}
\title{Visualizing severity level against the fluid levels, cartilage thickness
or patient age.}
\usage{
makeBox(dataMade, fluidLevels, cartThick, sevLevels, age)
}
\arguments{
\item{dataMade}{A list created by separateCSV.}

\item{fluidLevels}{A character variable that represents the fluid levels in
the data set}

\item{cartThick}{A character variable that represents the cartlidge thickness
in the data set}

\item{sevLevels}{A character variable that represents the severity level in
the data set}

\item{age}{A character variable that represents the age in the data set}
}
\value{
Returns a 3 box plots showcasing the distribution of fluid levels,
cartilage thickness and patient age
}
\description{
The function will take in a data set and a continuous variable
(i.e fluid levels, cartilage thickness or patient age). Then the data set
will be grouped based on the severity levels and 5 box plots will be created
to examine the distribution of the parameter variable across the severity
levels. The purpose of this function is to examine the distribution of the
variable across the severity levels.
}
\examples{

exampleDataPath <- system.file("extdata",
                               "example_data.csv",
                               package = "OsteoAnalizer")

dataList <- OsteoAnalizer::CleanCSVFile(filepath = exampleDataPath,
                               fluidLevels = "fluid_levels",
                               cartThick = "cartilage_thickness",
                               sevLevels = "severity_levels",
                               age = "patient_age")
OsteoAnalizer::makeBox(dataList, fluidLevels = "fluid_levels",
                               cartThick = "cartilage_thickness",
                               sevLevels = "severity_levels",
                               age = "patient_age")

}
\references{
Chatgpt. ChatGPT. (n.d.). https://openai.com/chatgpt

Holtz, Y. (n.d.). Control Ggplot2 Boxplot colors. – the R Graph Gallery.
https://r-graph-gallery.com/264-control-ggplot2-boxplot-colors.html

Li, F. (2022) sexDisaggregate:To Obtain and Analyze Sex-disaggregated Datasets,
Unpublished. URL https://github.com/lifangy6/sexDisaggregate"

R box plot. DataMentor. (n.d.). https://www.datamentor.io/r-programming/box-plot

Wickham H (2016). ggplot2: Elegant Graphics for Data Analysis.
}
\author{
{Krutika Joshi, \email{krutikadilip.joshi@mail.utoronto.ca}}
}
