% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{OA_data}
\alias{OA_data}
\title{Example dataset: Mock data set with Osteoarthritis Patient Information}
\format{
A data frame with 36 rows and 4 columns.
}
\source{
U.S. Department of Health and Human Services
}
\usage{
OA_data
}
\description{
This dataset, `OA_data`, contains information about patients with Osteoarthritis.
The variables present in this data set are the fluid_levels,
cartilage_thickness, severity_levels and the patient_age. The fluid_levels represent
the synovial fluid found in knee joints. This is measured in milliliters. The
cartilage_thickness variable as described by the name is the thickness of the
cartilage found in the knee and is measured in millimeters.
The severity_levels goes from a scale from 0 (least severe) to 4
(most severe). This variable represents the progression of osteoarthritis
in that individual. And lastly, the patient_age variable depicts the age of
the individual. This mock dataset is created to showcase the use of
osteoarthritis data in conjunction to the analysis functions
present in this project. The data was developed with reference to the data
set collected from U.S. Department of Health and Human Services.
}
\examples{
\dontrun{
 OA_data
}
}
\keyword{datasets}
